+++ src/factoriesAny.js
@@ -95,7 +95,7 @@ export { createZeros } from './function/matrix/zeros.js'
 export { createFft } from './function/matrix/fft.js'
 export { createIfft } from './function/matrix/ifft.js'
 export { createErf } from './function/special/erf.js'
-export { createRiemannZeta } from './function/special/riemannZeta.js'
+export { createZeta } from './function/special/zeta.js'
 export { createMode } from './function/statistics/mode.js'
 export { createProd } from './function/statistics/prod.js'
 export { createFormat } from './function/string/format.js'
