+++ src/function/special/zeta.js
@@ -1,9 +1,9 @@
 import { factory } from '../../utils/factory.js'
 
 const name = 'zeta'
-const dependencies = ['typed', 'config', 'multiply', 'pow', 'divide', 'factorial', 'equal', 'gamma', 'sin', 'subtract', 'add', 'Complex', 'BigNumber']
+const dependencies = ['typed', 'config', 'multiply', 'pow', 'divide', 'factorial', 'equal', 'gamma', 'sin', 'subtract', 'add', 'Complex', 'BigNumber', 'pi']
 
-export const createRiemannZeta = /* #__PURE__ */ factory(name, dependencies, ({ typed, config, multiply, pow, divide, factorial, equal, gamma, sin, subtract, add, Complex, BigNumber }) => {
+export const createZeta = /* #__PURE__ */ factory(name, dependencies, ({ typed, config, multiply, pow, divide, factorial, equal, gamma, sin, subtract, add, Complex, BigNumber, pi }) => {
   /**
    * Compute the Riemann Zeta function of a value using an infinite series for
    * all of the complex plane using Riemann's Functional equation.
@@ -25,7 +25,7 @@ export const createRiemannZeta = /* #__PURE__ */ factory(name, dependencies, ({ 
    *    math.zeta(math.i)  // returns 0.0033002..., -0.4181554491...i
    *
    *
-   * @param {number | Complex | BigNumber} s   A Real, Complex or BigNumber
+   * @param {number | Complex | BigNumber} s   A Real, Complex or BigNumber parameter to to the Riemann Zeta Function
    * @return {number | Complex | BigNumber}    The Riemann Zeta of `s`
    */
   function zeta (s) {
@@ -45,7 +45,7 @@ export const createRiemannZeta = /* #__PURE__ */ factory(name, dependencies, ({ 
     const n = Math.round(1.3 * 15 + 0.9 * Math.abs(s.im))
     if (s.re > -(n - 1) / 2) { return f(s, n) }
 
-    // Function Equation for reflection to re(s) < 1
+    // Functional Equation for reflection to re(s) < -( n - 1) / 2
     let c = multiply(pow(2, s), pow(Math.PI, subtract(s, 1)))
     c = multiply(c, (sin(multiply(Math.PI / 2, s))))
     c = multiply(c, gamma(subtract(1, s)))
@@ -70,8 +70,8 @@ export const createRiemannZeta = /* #__PURE__ */ factory(name, dependencies, ({ 
     if (x > -(n - 1) / 2) { return fBigNumber(x, n) }
 
     // Function Equation for reflection to x < 1
-    let c = multiply(pow(2, x), pow(BigNumber(Math.PI), subtract(x, 1)))
-    c = multiply(c, (sin(multiply(BigNumber(Math.PI / 2), x))))
+    let c = multiply(pow(2, x), pow(new BigNumber(pi), subtract(x, 1)))
+    c = multiply(c, (sin(multiply(new BigNumber(pi / 2), x))))
     c = multiply(c, gamma(subtract(1, x)))
     return multiply(c, zetaBigNumber(subtract(1, x)))
   }
@@ -92,10 +92,10 @@ export const createRiemannZeta = /* #__PURE__ */ factory(name, dependencies, ({ 
   }
   // Big Number alias
   function dBigNumber (k, n) {
-    let S = BigNumber(0)
-    const bn = BigNumber(n)
+    let S = new BigNumber(0)
+    const bn = new BigNumber(n)
     for (let j = k; j <= n; j++) {
-      const bj = BigNumber(j)
+      const bj = new BigNumber(j)
       const factor = divide(multiply(factorial(add(bn, subtract(bj, 1))), pow(4, bj)), multiply(factorial(subtract(bn, bj)), factorial(multiply(2, bj))))
       S = add(S, factor)
     }
@@ -106,7 +106,7 @@ export const createRiemannZeta = /* #__PURE__ */ factory(name, dependencies, ({ 
    * Calculate the positive Riemann Zeta function
    * @param {number} s   a real or complex number with s.re > 1
    * @param {number} n   a positive integer
-   * @return {number}    Zeta of s
+   * @return {number}    Riemann Zeta of s
    **/
   function f (s, n) {
     const c = divide(1, multiply(d(0, n), subtract(1, pow(2, subtract(1, s)))))
